25-05

+++ Administrator/views.py
@@ -2,7 +2,10 @@ from django.shortcuts import render,redirect
 from Administrator.models import*
 from Guest.models import*
 from User.models import*
+from Scrapcenter.models import*
 def District(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     districtdata=tbl_district.objects.all()
     if request.method=="POST":
         districtname=request.POST.get("txt_district")
@@ -13,11 +16,15 @@ def District(request):
         return render(request,'Administrator/District.html',{'districtdata':districtdata})
 
 def deldistrict(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_district.objects.get(id=did).delete()
     # return redirect('Administrator:District')
     return render(request,'Administrator/District.html',{"msg":"Data Deleted"})
 
 def editdistrict(request,eid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     editdata=tbl_district.objects.get(id=eid)
     if request.method=="POST":
         name=request.POST.get("txt_district")
@@ -28,6 +35,8 @@ def editdistrict(request,eid):
         return render(request,'Administrator/District.html',{'editdata':editdata})
 
 def AdminRegistration(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     admindata=tbl_admin.objects.all()
     if request.method=="POST":
         adminname=request.POST.get("txt_name")
@@ -44,10 +53,14 @@ def AdminRegistration(request):
         return render(request,'Administrator/AdminRegistration.html',{'admindata':admindata})
 
 def deladmin(request,aid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_admin.objects.get(id=aid).delete()
     return render(request,'Administrator/AdminRegistration.html',{"msg":"data deleted"})
 
 def editadmin(request,eid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     editdata=tbl_admin.objects.get(id=eid)
     if request.method=="POST":
         name=request.POST.get("txt_name")
@@ -64,6 +77,8 @@ def editadmin(request,eid):
         return render(request,'Administrator/AdminRegistration.html',{'editdata':editdata})
 
 def Category(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     categorydata=tbl_category.objects.all()
     if request.method=="POST":
         categoryname=request.POST.get("txt_category")
@@ -74,10 +89,14 @@ def Category(request):
         return render(request,'Administrator/Category.html',{'categorydata':categorydata})
 
 def delcategory(request,cid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_category.objects.get(id=cid).delete()
     return render(request,'Administrator/Category.html',{"msg":"data deleted"})
 
 def editcategory(request,eid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     editdata=tbl_category.objects.get(id=eid)
     if request.method=="POST":
         name=request.POST.get("txt_category")
@@ -88,6 +107,8 @@ def editcategory(request,eid):
         return render(request,'Administrator/Category.html',{'editdata':editdata})
 
 def Place(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     placedata=tbl_place.objects.all()
     districtdata=tbl_district.objects.all()
     if request.method=="POST":
@@ -104,10 +125,14 @@ def Place(request):
         return render(request,'Administrator/Place.html',{'districts':districtdata,'places':placedata})
 
 def delplace(request,pid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_place.objects.get(id=pid).delete()
     return render(request,'Administrator/Place.html',{"msg":"Data deleted"})
 
 def editplace(request,pid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     district=tbl_district.objects.all()
     editdata=tbl_place.objects.get(id=pid)
     if request.method=="POST":
@@ -122,6 +147,8 @@ def editplace(request,pid):
         return render(request,'Administrator/Place.html',{'editdata':editdata,'district':district})
 
 def SubCategory(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     subdata=tbl_subcategory.objects.all()
     categorydata=tbl_category.objects.all()
     if request.method=="POST":
@@ -136,10 +163,14 @@ def SubCategory(request):
         return render(request,'Administrator/SubCategory.html',{'categories':categorydata,'subcategories':subdata})
 
 def delsub(request,sid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_subcategory.objects.get(id=sid).delete()
     return render(request,'Administrator/SubCategory.html',{"msg":"Data deleted"})
 
 def editsub(request,sid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     category=tbl_category.objects.all()
     editdata=tbl_subcategory.objects.get(id=sid)
     if request.method=="POST":
@@ -152,8 +183,12 @@ def editsub(request,sid):
         return render(request,'Administrator/SubCategory.html',{'editdata':editdata,'categories':category})
 
 def Product(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     return render(request,'Administrator/Product.html')
 def Brand(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     branddata=tbl_brand.objects.all()
     categorydata=tbl_category.objects.all()
     if request.method=="POST":
@@ -168,10 +203,14 @@ def Brand(request):
         return render(request,'Administrator/Brand.html',{'branddata':branddata,'categories':categorydata})
 
 def delbrand(request,bid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_brand.objects.get(id=bid).delete()
     return render(request,'Administrator/Brand.html',{"msg":"Data deleted"})
 
 def editbrand(request,eid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     editdata=tbl_brand.objects.get(id=eid)
     category=tbl_category.objects.all()
     if request.method=="POST":
@@ -183,6 +222,8 @@ def editbrand(request,eid):
         return render(request,'Administrator/Brand.html',{'editdata':editdata,'categories':category})
 
 def Type(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     typedata=tbl_type.objects.all()
     if request.method=="POST":
         typename=request.POST.get("txt_type")
@@ -194,10 +235,14 @@ def Type(request):
         return render(request,'Administrator/Type.html',{'typedata':typedata})
 
 def deltype(request,tid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_type.objects.get(id=tid).delete()
     return render(request,'Administrator/Type.html',{"msg":"Data deleted"})
 
 def edittype(request,eid):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     editdata=tbl_type.objects.get(id=eid)
     if request.method=="POST":
         name=request.POST.get("txt_type")
@@ -208,19 +253,37 @@ def edittype(request,eid):
         return render(request,'Administrator/Type.html',{'editdata':editdata})
 
 def UserList(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     userdata=tbl_user.objects.all()
     return render(request,'Administrator/UserList.html',{'userdata':userdata})
 
 def Home(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     admin=request.session["aname"]
-    return render(request,'Administrator/Home.html',{'name':admin})
-
+    vehicle_count = tbl_addvehicle.objects.count()
+    user_count = tbl_user.objects.count()
+    scrap_count = tbl_scrapcenter.objects.count()
+    feedback_count = tbl_feedback.objects.count()
+    return render(request, "Administrator/Home.html", {
+        'vehicle_count': vehicle_count,
+        'user_count': user_count,
+        'scrap_count': scrap_count,
+        'feedback_count': feedback_count,
+        'name':admin
+    })
+   
 def ViewComplaint(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     ucdata=tbl_complaint.objects.filter(user__isnull=False)
     scdata=tbl_complaint.objects.filter(scrap__isnull=False)
     return render(request,'Administrator/ViewComplaint.html',{'ucdata':ucdata,'scdata':scdata})
 
 def Reply(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     cmptdata=tbl_complaint.objects.get(id=did)
     if request.method=="POST":
         reply_data=request.POST.get("txt_reply")
@@ -232,26 +295,36 @@ def Reply(request,did):
         return render(request,'Administrator/Reply.html')
 
 def ViewFeedback(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     vfdata=tbl_feedback.objects.all()
     return render(request,'Administrator/ViewFeedback.html',{'vfdata':vfdata})
 
 def ScrapcenterList(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     scdata=tbl_scrapcenter.objects.all()
     return render(request,'Administrator/ScrapcenterList.html',{'scdata':scdata})
 
 def accept(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     data=tbl_scrapcenter.objects.get(id=did)
     data.scrapcenter_status=1
     data.save()
     return render(request,'Administrator/ScrapcenterList.html',{"msg":"Verified"})
 
 def reject(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     data=tbl_scrapcenter.objects.get(id=did)
     data.scrapcenter_status=2
     data.save()
     return render(request,'Administrator/ScrapcenterList.html',{"msg":"Rejected"})
 
 def Model(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     modeldata=tbl_model.objects.all()
     category=tbl_category.objects.all()
     branddata=tbl_brand.objects.all()
@@ -267,10 +340,14 @@ def Model(request):
         return render(request,'Administrator/Model.html',{'modeldata':modeldata,'branddata':branddata,'category':category})
 
 def delmodel(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_model.objects.get(id=did).delete()
     return render(request,'Administrator/Model.html',{"msg":"Data deleted"})
 
 def editmodel(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     editdata=tbl_model.objects.get(id=did)
     editcategory=editdata.brand.category.id
     category=tbl_category.objects.all()
@@ -290,6 +367,8 @@ def AjaxBrand(request):
     branddata=tbl_brand.objects.filter(category=category)
     return render(request,"Administrator/AjaxBrand.html",{'brand':branddata})
 def ScrapRate(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     category=tbl_category.objects.all()
     ratedata=tbl_scraprate.objects.all()
     if request.method=="POST":
@@ -303,9 +382,13 @@ def ScrapRate(request):
     else:
         return render(request,'Administrator/ScrapRate.html',{'ratedata':ratedata,'category':category})
 def delrate(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     tbl_scraprate.objects.get(id=did).delete()
     return render(request,'Administrator/ScrapRate.html',{"msg":"Data deleted"})
 def editrate(request,did):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
     category=tbl_category.objects.all()
     editdata=tbl_scraprate.objects.get(id=did)
     if request.method=="POST":
@@ -316,3 +399,12 @@ def editrate(request,did):
         return render(request,'Administrator/ScrapRate.html',{"msg":"Data updated"})
     else:
         return render(request,'Administrator/ScrapRate.html',{'editdata':editdata,'category':category})
+
+def Logout(request):
+    if 'aid' not in request.session:
+        return redirect('Guest:Login')
+    del request.session['aid']
+    return redirect('Guest:Login')
+
+
+
